[config_validation] Fold validator tests into existing test file
Merge the validator coverage tests from the separate
test_config_validation_validators.py into test_config_validation.py and
delete the standalone file, keeping all config_validation tests in one
place. The existing tests are renamed to the canonical cv. alias to match.
